add basic positional animation
The logic here is to use the JavaFX Layout system (which is complicated
and I don't want to have to manually extend Region if i can avoid it for
a prototype) but also to animate changes in the positions of the nodes.

To accomplish this, I've used an anchorpane called DisplayOverlay
stacked on top of a basic layout system (at the moment the whole process
is an HBox and each stage of the process is a VBox). Changes to the
positions of the preparation nodes are not automatically pushed to the
displayoverlay (not bound), the logic being that the program will choose
when to reconcile the displays.

The stackpane is wrapped in a scrollpane for long processes, but I
haven't done anything with that yet.